+++ test/Bgetre_comment.c
@@ -6,7 +6,7 @@
 #include <string.h>
 #include <unistd.h>
 
-extern int Bgetre(Biobuf *bp, Reprog *progp, Resub *mp, int msize, char **wp, size_t *wsize);
+extern size_t Bgetre(Biobuf *bp, Reprog *progp, Resub *mp, int msize, char **wp, size_t *wsize);
 
 #define C_COMMENT "/\\*.*\\*/"
